+++ Models/TrafficLight.cs
@@ -12,9 +12,9 @@ namespace Server.Models
         public string hash_code {get; set;}
 
         [Column("long")]
-        public float long_ {get; set;}
+        public double long_ {get; set;}
 
-        public float lat {get; set;}
+        public double lat {get; set;}
 
         public int district_id {get; set;}
         [ForeignKey("district_id")]
